pyproject.toml
# if poetry update takes forever
# export PYTHON_KEYRING_BACKEND=keyring.backends.fail.Keyring

[tool.poetry]
name = "webtoolkit"
version = "0.0.184"
description = "Web tools and interfaces for Internet data processing."
authors = ["Iwan Grozny <renegat@renegat0x0.ddns.net>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"
tldextract = "^5.1.2"
beautifulsoup4 = "^4.13.5"
lxml = "^5.4.0"
brutefeedparser="^0.10.5"
pytz = "^2024.2"
psutil="*"
url-cleaner="*"
#cleanurl="*"  TODO may change url structure
ua-generator="^2.0.17"

# Enable and crawlers as needed

# [Requests]
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
